requirements.txt
torch~=1.7.0+cpu
numpy~=1.19.4
torchvision~=0.8.1+cpu
Pillow~=8.0.1
matplotlib~=3.3.3
omegaconf~=2.0.5
thop~=0.0.31.post2005241907
wandb~=0.10.12